Refactor: Authorizer abstracts, namespace fixes
