+++ BLM.EF6/EfContextInfo.cs
@@ -2,6 +2,7 @@
 using System.Linq;
 using System.Security.Principal;
 using System.Threading.Tasks;
+using BLM.Interfaces;
 
 namespace BLM.EF6
 {
