+++ BLM.Tests/ListenerTests.cs
@@ -3,6 +3,7 @@ using System.Threading;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using BLM.Interfaces;
 using BLM.Interfaces.Listen;
 
 #pragma warning disable 1998
