+++ BLM.Tests/MockCollectionAuthorizer2.cs
@@ -1,6 +1,7 @@
 using System.Linq;
 using BLM.Interfaces.Authorize;
 using System.Threading.Tasks;
+using BLM.Interfaces;
 
 namespace BLM.Tests
 {
