+++ BLM.Tests/MockCreateAuthorizer.cs
@@ -1,5 +1,6 @@
 ﻿using BLM;
 using System.Threading.Tasks;
+using BLM.Interfaces;
 
 namespace BLM.Tests
 {
