+++ BLM.Tests/MockRemoveAuthorizer.cs
@@ -1,4 +1,5 @@
 using System.Threading.Tasks;
+using BLM.Interfaces;
 
 namespace BLM.Tests
 {
