+++ BLM/Authorize.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using BLM.Interfaces;
 using BLM.Interfaces.Authorize;
 
 namespace BLM
