+++ BLM/AuthorizeModify.cs
@@ -1,4 +1,5 @@
 ﻿using System.Threading.Tasks;
+using BLM.Interfaces;
 using BLM.Interfaces.Authorize;
 
 namespace BLM
