+++ BLM/Interfaces/IRepository.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Security.Principal;
 using System.Threading.Tasks;
 
-namespace BLM
+namespace BLM.Interfaces
 {
 
     public interface IRepository
